requirements.txt
instructor>=0.4.8
openai>=1.3.7
requests>=2.31.0
pillow>=10.1.0
python-dotenv>=1.0.0
backoff>=2.2.1
pytest-asyncio>=0.23.2
aiohttp>=3.9.1
